smart lms W deploy/requirements.txt
Flask
Flask-PyMongo
python-dotenv
google-generativeai
gunicorn
requests
werkzeug
pytz
PyMuPDF
dnspython
Flask-Cors
torch
transformers
sentencepiece
sentence-transformers
chromadb
